🎨 HistoryHeader 분리
