+++ client/src/components/molecules/PlayController/style.tsx
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const Wrapper = styled.div`
+  ${({ theme }) => theme.flex.row}
+  align-self: center
+`;
