+++ client/src/hooks/useLinkClick/index.tsx
@@ -1,4 +1,5 @@
 import useProjectId from 'hooks/useRoomId';
+import { useCallback } from 'react';
 import { useHistory } from 'react-router';
 import { TPage } from 'types/page';
 
@@ -6,9 +7,9 @@ const useLinkClick = (path: TPage) => {
   const history = useHistory();
   const projectId = useProjectId();
 
-  const handleLinkClick = () => {
+  const handleLinkClick = useCallback(() => {
     history.push(`/${path}/${projectId}`);
-  };
+  }, [history, projectId]);
 
   return handleLinkClick;
 };
